+++ packages/deploy-tooling/src/index.ts
@@ -1,0 +1,4 @@
+import task from './ui5';
+export * from './cli';
+
+export { task };