+++ packages/axios-extension/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/axios-extension
 
+## 1.0.3
+
+### Patch Changes
+
+-   7fd2810: improved logging
+
+## 1.0.2
+
+### Patch Changes
+
+-   8e059ae: consider segment parameters in service uri
+
 ## 1.0.1
 
 ### Patch Changes